Configuration should be verbose
Previously, it was using an array to hold properties. This made it
difficult to work with and track where properties were being set and
get as well as what properties actually existed on the class. It had
no benefit over regular properties for both performance and development
so it's been changed in favour of actual defined properties.

As well, a bit of clarification on the existing retry option was added
through a change in naming. It should now properly reflect that it only
affects reads (and only a subset of reads).

Writes also have their own configuration options available for retry
behaviour to be added in the next few commits.