Retries should back off after failed attempts
Previously retries just happened within the same process in a for loop.
If the process was fast enough, it stood a chance that the retry would
just encounter the same issue over and over again e.g. in the case of
connectivity issues.

This is the groundwork necessary to implement a working retry backoff.