Configuration data structure is used everywhere
With the expansion of the possible configuration options that are now
available, it's a lot easier to pass the data structure around. While
this causes a lot more coupling between the classes and generally mucks
things up, it's the nature of configuration and utility classes in that
generally everything needs them.