Nuanced retry behaviour for each write operation
Retries for writes are broken down into three distinct operations:
singular inserts, batch inserts, and updates.

All operations have to be idempotent in order to be retried: if retrying
them multiple times could cause multiple documents to be inserted or
fields to be added or removed each time, we shouldn't retry at the
wrapper level. Applications making use of the wrapper will have to
determine the best course of action in the case of a failure for
non-idempotent operations.

In the case of both singular and batch inserts, this means we have to
generate document id(s) client-side before we send over the command.

A bit of special logic is in place to not generate an id of one already
exists (in case implementing applications have their own id types they'd
want to use).

The idea behind single insert retry logic is that if a connection
exception occurs on any attempt that we should retry in case the command
never went through / executed / executed but we don't know about it /
etc.

We're guaranteed idempotent inserts and can even detect it by catching
duplicate key errors and parsing out the index it failed on, only
ignoring duplicate key exceptions that occur on duplicate ids past the
first attempt.

Sadly, return values have to be hard-coded and kept in sync with the
underlying driver since we can't get a mock successful response on
a failed-but-fine case.

Updates are a bit tougher to ensure idempotency. The gist of it is that
we have to manually filter out update queries that aren't idempotent
by the operations used. Operations like $push and $bit mean we can't
reliably retry the query and should only attempt it once.

Batch inserts are much looser than the other two write retries due to
how little error handling we can reasonably perform without affecting
performance. We have to ignore all errors on subsequent retries even if
we only want to ignore duplicate key errors since the entire batch
insert stops if any error is encountered; the only way to continue on
an error is also the only way to ignore a specific (in this case all)
exceptions.